Extract and reuse export test Scenario to its own file
