+++ test/java/org/opendatakit/briefcase/export/CsvSubmissionMappersEncodeValueTest.java
@@ -17,9 +17,9 @@
 package org.opendatakit.briefcase.export;
 
 import static org.junit.Assert.assertThat;
-import static org.opendatakit.briefcase.export.CsvFieldMappersTest.Scenario.createField;
 import static org.opendatakit.briefcase.export.CsvSubmissionMappers.encodeMainValue;
 import static org.opendatakit.briefcase.export.CsvSubmissionMappers.encodeRepeatValue;
+import static org.opendatakit.briefcase.export.Scenario.createField;
 
 import org.hamcrest.Matchers;
 import org.javarosa.core.model.DataType;
